+++ src/nodes/server.rs
@@ -213,10 +213,10 @@ impl server {
 
     /// Run the server to process incoming packets and handle fragment assembly
     pub fn run(&mut self) {
-        // Initialize the logger
+        /*// Initialize the logger
         if let Err(e) = env_logger::builder().try_init() {
             eprintln!("Failed to initialize logger: {}", e);
-        }
+        }*/
 
         info!("Server {} started running.", self.id);
         loop {
