--- src/utils/mod.rs
@@ -1,2 +1,0 @@
-mod config;
-pub(crate) mod serialization_fix;
